requirements.txt
asgiref==3.7.2
Django==4.2.5
psycopg2-binary==2.9.7
sqlparse==0.4.4
celery==5.3.4
redis==5.0.0
django-froala-editor==4.1.3
Pillow==10.1.0